formatting, minor refactoring
